完成所有功能

+++ services/users.js
@@ -184,23 +184,82 @@ module.exports.userId = function(req, res) {
   			   password : '123456',
   			   database : 'open_data'
   			 });
-  			 connection.connect();
-  			 connection.query(`DELETE FROM users WHERE id IN (${id})`, function(error, results, fields) {
-  				 if (error) throw res.status(200).json({success:0,message: '删除失败'});
-  				 	// if (error) throw error;
-  				 	return res.status(200).json({
-  				 		success:200,
-  				 		message: '删除成功'
-  				 	});
-  				 
-  			 })
+  			   //连接数据库
+  			connection.connect();
+			connection.query(`SELECT type FROM users WHERE  id=${id}`, function(error, results, fields) {
+				if (results[0].type == 0) {
+					var connection = mysql.createConnection({
+					   host:'localhost',
+					   user: 'root',
+					   password : '123456',
+					   database : 'open_data'
+					 });
+					   //连接数据库
+					connection.connect();
+					connection.query(`SELECT * FROM users WHERE  type=0`, function(error, results, fields) {
+						if (results.length == 1) {
+							return res.status(200).json({
+								success:0,
+								message: '系统中至少要有一个管理员账户'
+							});
+						} else {
+							var connection = mysql.createConnection({
+							   host:'localhost',
+							   user: 'root',
+							   password : '123456',
+							   database : 'open_data'
+							 });
+							 connection.connect();
+							 connection.query(`DELETE FROM users WHERE id IN (${id})`, function(error, results, fields) {
+								 if (error) throw res.status(200).json({success:0,message: '删除失败'});
+								 	// if (error) throw error;
+								 	return res.status(200).json({
+								 		success:200,
+								 		message: '删除成功'
+								 	});
+								 
+							 })
+						}
+					})
+				} else {
+					var connection = mysql.createConnection({
+					   host:'localhost',
+					   user: 'root',
+					   password : '123456',
+					   database : 'open_data'
+					 });
+					 connection.connect();
+					 connection.query(`DELETE FROM users WHERE id IN (${id})`, function(error, results, fields) {
+						 if (error) throw res.status(200).json({success:0,message: '删除失败'});
+						 	// if (error) throw error;
+						 	return res.status(200).json({
+						 		success:200,
+						 		message: '删除成功'
+						 	});
+						 
+					 })
+				}
+			})
   		}
   	})
   	connection.end();
   }
   
+   // 添加用户或修改用户信息
    module.exports.useredit = function(req, res) {
 	   const tokenKey = req.headers.authorization;
+	   const id = req.body.id;
+	   const employeeName = req.body.employeeName;
+	   const sex = Number(req.body.sex);
+	   const address = req.body.address;
+	   const birthday = req.body.birthday;
+	   const department = req.body.department;
+	   const job = req.body.job;
+	   const employeeID = Number(req.body.employeeID);
+	   const username = req.body.username;
+	   const password = req.body.password;
+	   const type = Number(req.body.type);
+	   const openr = req.body.openr;
 	   var connection = mysql.createConnection({
 	      host:'localhost',
 	      user: 'root',
@@ -217,8 +276,44 @@ module.exports.userId = function(req, res) {
 		   		   message:'非法操作'
 		   		 });
 		   } else {
-			   console.log(req.body);
+			   if (id == null) { // 添加新用户
+				   var connection = mysql.createConnection({
+				      host:'localhost',
+				      user: 'root',
+				      password : '123456',
+				      database : 'open_data'
+				    });
+				      //连接数据库
+				   connection.connect();
+				   connection.query(`INSERT INTO users (employeeName,sex,address,birthday,department,job,employeeID,username,password,type,openr) VALUES ('${employeeName}',${sex},'${address}',${birthday},'${department}','${job}','${employeeID}','${username}','${password}',${type},'${openr}')`, function(error, results, fields) {
+					   if (error) throw res.status(200).json({success:0,message: '操作失败'});
+					   return res.status(200).json({
+					   	success:200,
+					   	message: '添加成功'
+					   });
+				   })
+			   } else { // 修改用户信息
+				   var connection = mysql.createConnection({
+				      host:'localhost',
+				      user: 'root',
+				      password : '123456',
+				      database : 'open_data'
+				   });
+				   connection.connect();
+				   connection.query(`UPDATE users SET employeeName='${employeeName}',sex=${sex},address='${address}',birthday=${birthday},department='${department}',job='${job}',employeeID='${employeeID}',username='${username}',password='${password}',type=${type},openr='${openr}' WHERE id=${id}`, function(error, results, fields) {
+				   	if (error) throw res.status(200).json({success:0,message: '操作失败'});
+				   	// if (error) throw error;
+				   	return res.status(200).json({
+				   		success:200,
+				   		message: '修改成功'
+				   	});
+				   })
+			   }
 		   }
 	   })
 	   connection.end();
    }
+   
+   
+
+
